move ast tests into tests/
