+++ core/src/ast/mod.rs
@@ -12,8 +12,6 @@ use std::fmt;
 mod def;
 mod expr;
 mod query;
-#[cfg(test)]
-mod test;
 
 pub use def::{DateMatch, DatePattern, Def, DefEntry, Defs, ExprString, Property};
 pub use expr::{Expr, Precedence};
@@ -221,58 +219,5 @@ impl fmt::Display for DateToken {
             DateToken::Plus => write!(fmt, "+"),
             DateToken::Error(ref e) => write!(fmt, "<{}>", e),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::ast::DateToken;
-
-    use super::Function;
-
-    const FUNCTIONS: &'static [Function] = &[
-        Function::Sqrt,
-        Function::Exp,
-        Function::Ln,
-        Function::Log2,
-        Function::Log10,
-        Function::Sin,
-        Function::Cos,
-        Function::Tan,
-        Function::Asin,
-        Function::Acos,
-        Function::Atan,
-        Function::Sinh,
-        Function::Cosh,
-        Function::Tanh,
-        Function::Asinh,
-        Function::Acosh,
-        Function::Atanh,
-        Function::Log,
-        Function::Hypot,
-        Function::Atan2,
-        Function::Fac,
-    ];
-
-    #[test]
-    fn roundtrip_function_names() {
-        for &func in FUNCTIONS {
-            assert_eq!(Function::from_name(func.name()), Some(func));
-        }
-    }
-
-    #[test]
-    fn date_tokens_display() {
-        assert_eq!(DateToken::Colon.to_string(), ":");
-        assert_eq!(DateToken::Dash.to_string(), "-");
-        assert_eq!(DateToken::Space.to_string(), " ");
-        assert_eq!(DateToken::Plus.to_string(), "+");
-        assert_eq!(DateToken::Literal("a".to_owned()).to_string(), "a");
-        assert_eq!(DateToken::Number("123".to_owned(), None).to_string(), "123");
-        assert_eq!(
-            DateToken::Number("123".to_owned(), Some("456".to_owned())).to_string(),
-            "123.456"
-        );
-        assert_eq!(DateToken::Error("test".to_owned()).to_string(), "<test>");
     }
 }
